# src/main/gen/src/main/java/io/security/portfolio/api/AuditLogsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package io.security.portfolio.api;

import org.springframework.format.annotation.DateTimeFormat;
import io.security.portfolio.dto.ListAuditLogs200Response;
import java.time.OffsetDateTime;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-07T12:21:57.339446635+09:00[Asia/Seoul]", comments = "Generator version: 7.7.0")
@Validated
@Tag(name = "AuditLogs", description = "the AuditLogs API")
public interface AuditLogsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /api/audit/logs : 감사 로그 조회
     *
     * @param actorId  (optional)
     * @param resourceType  (optional)
     * @param resourceId  (optional)
     * @param action  (optional)
     * @param decision  (optional)
     * @param from 조회 시작 시각 (optional)
     * @param to 조회 종료 시각 (optional)
     * @return 감사 로그 목록 (status code 200)
     */
    @Operation(
        operationId = "listAuditLogs",
        summary = "감사 로그 조회",
        tags = { "AuditLogs" },
        responses = {
            @ApiResponse(responseCode = "200", description = "감사 로그 목록", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ListAuditLogs200Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/audit/logs",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ListAuditLogs200Response> listAuditLogs(
        @Parameter(name = "actorId", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "actorId", required = false) Long actorId,
        @Parameter(name = "resourceType", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "resourceType", required = false) String resourceType,
        @Parameter(name = "resourceId", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "resourceId", required = false) Long resourceId,
        @Parameter(name = "action", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "action", required = false) String action,
        @Parameter(name = "decision", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "decision", required = false) String decision,
        @Parameter(name = "from", description = "조회 시작 시각", in = ParameterIn.QUERY) @Valid @RequestParam(value = "from", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) OffsetDateTime from,
        @Parameter(name = "to", description = "조회 종료 시각", in = ParameterIn.QUERY) @Valid @RequestParam(value = "to", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) OffsetDateTime to
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : { \"logs\" : [ { \"reason\" : \"reason\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"resourceId\" : 1, \"actorId\" : 6, \"decision\" : \"decision\", \"extra\" : \"{}\", \"action\" : \"action\", \"id\" : 0, \"resourceType\" : \"resourceType\" }, { \"reason\" : \"reason\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"resourceId\" : 1, \"actorId\" : 6, \"decision\" : \"decision\", \"extra\" : \"{}\", \"action\" : \"action\", \"id\" : 0, \"resourceType\" : \"resourceType\" } ] }, \"success\" : true, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}