# src/main/gen/src/main/java/io/security/portfolio/api/AuthApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package io.security.portfolio.api;

import io.security.portfolio.dto.Login200Response;
import io.security.portfolio.dto.LoginRequest;
import io.security.portfolio.dto.LogoutRequest;
import io.security.portfolio.dto.ModelApiResponse;
import io.security.portfolio.dto.RefreshTokenRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-07T12:21:57.339446635+09:00[Asia/Seoul]", comments = "Generator version: 7.7.0")
@Validated
@Tag(name = "Auth", description = "the Auth API")
public interface AuthApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /api/auth/login : 로그인 (JWT 발급)
     *
     * @param loginRequest  (required)
     * @return 로그인 성공 (status code 200)
     */
    @Operation(
        operationId = "login",
        summary = "로그인 (JWT 발급)",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "로그인 성공", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Login200Response.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/auth/login",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Login200Response> login(
        @Parameter(name = "LoginRequest", description = "", required = true) @Valid @RequestBody LoginRequest loginRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : { \"expiresIn\" : 0, \"accessToken\" : \"accessToken\", \"tokenType\" : \"Bearer\", \"refreshToken\" : \"refreshToken\" }, \"success\" : true, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/auth/logout : 로그아웃 (리프레시 토큰 폐기)
     *
     * @param logoutRequest  (required)
     * @return 로그아웃 성공 (status code 200)
     */
    @Operation(
        operationId = "logout",
        summary = "로그아웃 (리프레시 토큰 폐기)",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "로그아웃 성공", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ModelApiResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/auth/logout",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<ModelApiResponse> logout(
        @Parameter(name = "LogoutRequest", description = "", required = true) @Valid @RequestBody LogoutRequest logoutRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : \"\", \"success\" : true, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/auth/refresh : 액세스 토큰 재발급
     *
     * @param refreshTokenRequest  (required)
     * @return 토큰 재발급 성공 (status code 200)
     */
    @Operation(
        operationId = "refreshToken",
        summary = "액세스 토큰 재발급",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "토큰 재발급 성공", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Login200Response.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/auth/refresh",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Login200Response> refreshToken(
        @Parameter(name = "RefreshTokenRequest", description = "", required = true) @Valid @RequestBody RefreshTokenRequest refreshTokenRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : { \"expiresIn\" : 0, \"accessToken\" : \"accessToken\", \"tokenType\" : \"Bearer\", \"refreshToken\" : \"refreshToken\" }, \"success\" : true, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}