# src/main/java/io/security/portfolio/api/auth/TokenResponse.java
package io.security.portfolio.api.auth;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;

@Getter
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class TokenResponse {
    private String accessToken;
    private String refreshToken;
    private String tokenType;   // "Bearer"
    private Long expiresIn;     // seconds
}