# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Document Authorization API
  version: 1.0.0
  description: |
    RBAC + Team Role + ABAC + ACL(PBAC 기반) 구조를 가진 문서 인가 데모 API.
    - 글로벌 RBAC: users / roles / user_roles
    - 팀/조직: teams / team_members
    - 문서: documents
    - ACL(문서별 권한): document_permissions
    - 감사 로그: audit_logs
    - 인증: refresh_tokens
servers:
  - url: http://localhost:8080
    description: Local dev server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 요청 성공 여부
        message:
          type: string
          nullable: true
          description: 추가 메시지 (없으면 null)
        data:
          nullable: true
          description: 실제 응답 데이터 (엔드포인트별로 타입 상이)

    # =====================
    # Auth
    # =====================
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
          description: 액세스 토큰 만료까지 남은 시간(초)
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    LogoutRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string

    # =====================
    # User
    # =====================
...
    AuditLogResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        actorId:
          type: integer
          format: int64
        action:
          type: string
        resourceType:
          type: string
          description: "DOCUMENT, TEAM, USER 등"
        resourceId:
          type: integer
          format: int64
        decision:
          type: string
          description: "PERMIT, DENY"
        reason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        extra:
          type: object
          nullable: true
          description: "추가 컨텍스트/정책 평가 결과(JSON)"
    AuditLogListResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogResponse'
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: 로그인 (JWT 발급)
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TokenResponse'
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: 액세스 토큰 재발급
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: 토큰 재발급 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TokenResponse'
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: 로그아웃 (리프레시 토큰 폐기)
      operationId: logout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: 로그아웃 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # =====================
  # Users
  # =====================
...
  /api/audit/logs:
    get:
      tags: [AuditLogs]
      summary: 감사 로그 조회
      operationId: listAuditLogs
      security:
        - bearerAuth: []
      parameters:
        - name: actorId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: resourceType
          in: query
          required: false
          schema:
            type: string
        - name: resourceId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: action
          in: query
          required: false
          schema:
            type: string
        - name: decision
          in: query
          required: false
          schema:
            type: string
            description: "PERMIT, DENY"
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: 조회 시작 시각
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: 조회 종료 시각
      responses:
        '200':
          description: 감사 로그 목록
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuditLogListResponse'